Corrects counts in trace detail screen
While I don't currently agree with the counts being helpful. This fixes
the counts such that they add up. Before, they were count of unique span
ID, not actual spans. This is confusing..

More work to follow, but this is enough to review on its own.
