+++ zipkin-ui/js/component_data/trace.js
@@ -2,7 +2,7 @@ import {component} from 'flightjs';
 import $ from 'jquery';
 import {getError} from '../../js/component_ui/error';
 import traceToMustache from '../../js/component_ui/traceToMustache';
-import {correctForClockSkew} from '../skew';
+import {treeCorrectedForClockSkew} from '../skew';
 
 export function toContextualLogsUrl(logsUrl, traceId) {
   if (logsUrl) {
@@ -13,7 +13,7 @@ export function toContextualLogsUrl(logsUrl, traceId) {
 
 // Converts the response into data for trace.mustache. Missing required data will raise an error.
 export function convertSuccessResponse(rawResponse, logsUrl) {
-  const corrected = correctForClockSkew(rawResponse);
+  const corrected = treeCorrectedForClockSkew(rawResponse);
   const modelview = traceToMustache(corrected, logsUrl);
   return {modelview, trace: rawResponse};
 }
